add entity customer and implementation some fixes in templates cars-list carform

+++ src/main/java/com/DawidM/SavingDataCar/controllers/CarController.java
@@ -57,7 +57,7 @@ public class CarController {
     }
 
     @GetMapping("/delete")
-    public String delete(@RequestParam("carId") int id){
+    public String delete(@RequestParam("carId") Long id){
 
         carService.deleteById(id);
         return "redirect:/cars/list";
